backend/requirements.txt
django==1.10.4
djangorestframework==3.5.3
django-cors-headers==1.3.1
